+++ src/PhpSpreadsheet/Reader/Gnumeric.php
@@ -272,6 +272,11 @@ class Gnumeric extends BaseReader
             //        name in line with the formula, not the reverse
             $this->spreadsheet->getActiveSheet()->setTitle($worksheetName, false, false);
 
+            $visibility = $sheetOrNull->attributes()['Visibility'] ?? 'GNM_SHEET_VISIBILITY_VISIBLE';
+            if ((string) $visibility !== 'GNM_SHEET_VISIBILITY_VISIBLE') {
+                $this->spreadsheet->getActiveSheet()->setSheetState(Worksheet::SHEETSTATE_HIDDEN);
+            }
+
             if (!$this->readDataOnly) {
                 (new PageSetup($this->spreadsheet))
                     ->printInformation($sheet)